work around asterisk possibly not hearing all 5 tones of a quarter
since the tones are played rapidly and are going over the internet,
they might not get heard properly/separately by asterisk.  keep
track of the timing (using Time::HiRes) of tones received and if we
hear 3 in less than 0.5 seconds, consider that good enough to count
as a quarter and ignore the 4th and 5th ones we might still hear.